LDAP: move ldap modules to infra folder
Since those modules do not defined as services, it feels more
correct to place them in `infra` folder
