+++ pkg/infra/multildap/multildap.go
@@ -3,8 +3,8 @@ package multildap
 import (
 	"errors"
 
+	"github.com/grafana/grafana/pkg/infra/ldap"
 	"github.com/grafana/grafana/pkg/models"
-	"github.com/grafana/grafana/pkg/services/ldap"
 )
 
 // GetConfig gets LDAP config
