+++ pkg/login/auth.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 
 	"github.com/grafana/grafana/pkg/bus"
+	"github.com/grafana/grafana/pkg/infra/ldap"
 	"github.com/grafana/grafana/pkg/models"
-	"github.com/grafana/grafana/pkg/services/ldap"
 )
 
 var (
