+++ pkg/login/ldap_login.go
@@ -1,8 +1,8 @@
 package login
 
 import (
+	"github.com/grafana/grafana/pkg/infra/multildap"
 	"github.com/grafana/grafana/pkg/models"
-	"github.com/grafana/grafana/pkg/services/multildap"
 	"github.com/grafana/grafana/pkg/services/user"
 	"github.com/grafana/grafana/pkg/setting"
 	"github.com/grafana/grafana/pkg/util/errutil"
